Changed return type of indexCount to uint32_t and now cast size_t to uint32_t

+++ DVR/src/unit_cube.h
@@ -14,7 +14,7 @@ public:
 	inline void bind() const { m_vao.bind(); }
 	inline void unbind() const { m_vao.unbind(); }
 
-	inline int indexCount() const { return m_indices.size(); }
+	inline uint32_t indexCount() const { return uint32_t(m_indices.size()); }
 
 private:
